Update HangmanGame Class.py
Rearranged order of the methods to better follow the flow of the game, added guessed_words attribute to track the score of player, added the add_player_score method which adds the players username and score to the leaderboard text file. 

See google docs for more details on the changes.