Update and rename HangmanGame Class.py to HangmanGame.py
Added reset methods for the sake of resetting values on the full game loop and guess loop. 

Made adjustments to the methods as well to fix some of the errors and typos. Added various minor functions to improve formatting and fix bugs with the program. Also, added attributes to track the reset, win, and hangman status to for the loops as well as for conditionals in the main function.